Refactor: Move auth subcommand to its own module for better structure
